limepcie: stop truncating the firmware version to 16 bits
ReadInfo composes the firmware id as (payload[9] << 16) | payload[0],
but info.firmware was a uint16_t, so the high byte was silently
dropped and the reported version was always just payload[0]. Widen
the field to uint32_t.

Fixes #310

+++ drivers/linux/limepcie/limepcie_core.c
@@ -68,7 +68,7 @@ struct limepcie_device_attributes {
 
 struct deviceInfo {
     uint64_t serialNumber;
-    uint16_t firmware;
+    uint32_t firmware;
     uint8_t boardId;
     uint8_t protocol;
     uint8_t hardwareVer;
